[REV] some fixes

+++ addons/account/models/account_analytic_line.py
@@ -123,8 +123,11 @@ class AccountAnalyticLine(models.Model):
         return res
 
     def unlink(self):
+        self.env.flush_all()
+        if not self:
+            return False
         affected_move_lines = self.move_line_id
         res = super().unlink()
-        self.env.invalidate_all()  # reset all caches as before
+        #self.env.invalidate_all()  # reset all caches as before
         affected_move_lines._update_analytic_distribution()
         return res
